Moved the shop from /tests to /tienda
