+++ app/tienda/carrito/page.tsx
